Fix deinit race condition
There was a race conditition in iolink_m_deinit where it could free
thread resources before the thread had exited. Fix by polling for task
exit.

Note that exiting the thread will leak some resources on at least
Linux. Fixing this requires extending the OSAL layer with a thread
destroy function or similar. For normal operation we don't expect the
thread to exit, however for the unit tests the thread is
created/destroyed for every test. This may not be required, it would
be better practice to write the unit tests to test the underlying
modules directly.