Set internal state before indicating to upper layers
In Linux or a multi-threaded enviroment, IOLINK_DL_EVENT_MH could be set twice, but only read once, resulting in a deadlock after the message handler transitioned from STARTUP_2 -> PREOPERATE_6. Changed the order of operations to ensure that the internal state is set before sending an indication of the state change to upper layers.